Moves notifications service calls already in RTK to separated files.
