+++ cardstack/src/services/hub/notifications/hub-notifications-types.ts
@@ -1,0 +1,3 @@
+export interface RegisterFCMTokenQueryParams {
+  fcmToken: string;
+}
